+++ packages/interfaces/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @graphprotocol/interfaces
 
+## 0.6.2
+
+### Patch Changes
+
+- Ensure ServiceRegistry loads the correct interface
+
+## 0.6.1
+
+### Patch Changes
+
+- Add ServiceRegistered to LegacyServiceRegistry interface
+
 ## 0.6.0
 
 ### Minor Changes